[BLINK-17750684] external shuffle support compression
Summary: fix T17750684

Reviewers: 淘江, kevin.cyj, 云骞

Reviewed By: 云骞

Subscribers: caizhi.wcz, 之信, P710480

Differential Revision: https://aone.alibaba-inc.com/code/D771440

+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/external/writer/AsynchronousPartitionedStreamFileReaderDelegate.java
@@ -49,7 +49,7 @@ public class AsynchronousPartitionedStreamFileReaderDelegate implements RequestD
 	private final List<PartitionIndex> partitionIndices;
 
 	private int nextPartitionIdx;
-	private int nextOffset;
+	private long nextOffset;
 
 	public AsynchronousPartitionedStreamFileReaderDelegate(IOManager ioManager, FileIOChannel.ID channel,
 														   List<MemorySegment> segments,
